implemented utxout model, coinbase/normal txes, block rewards, miner fees, mempool, tx & mempool validation. Extensive refactorings, unit tests

+++ ArakCoin/Utilities.cs
@@ -18,9 +18,9 @@ public class Utilities
 	/**
 	 * Returns the current UTC time since epoch (in seconds)
 	 */
-	public static int getTimestamp()
+	public static long getTimestamp()
 	{
-		return (int)DateTime.UtcNow.Subtract(DateTime.UnixEpoch).TotalSeconds;
+		return (long)DateTime.UtcNow.Subtract(DateTime.UnixEpoch).TotalSeconds;
 	}
 
 	/**
